define lib.rs, use regex for file name matching
