+++ Cargo.toml
@@ -6,3 +6,4 @@ edition = "2018"
 
 [dependencies]
 tokio = "0.1.20"
+regex = "1"
