+++ src/main.rs
@@ -5,11 +5,14 @@ use std::path::Path;
 use std::fs::{File};
 use std::io::Read;
 
-mod findr;
+mod lib;
 
+// Find a file {{f}} in directory {{d}} and print it's contents:
+// cargo run {{d}} {{f}}
 fn main() {
     let args: Vec<String> = env::args().collect();
-    match findr::visit_dirs(Path::new(&args[1]), &args[2], &|d| {
+    match lib::find(Path::new(&args[1]), &args[2], &|d| {
+        println!("File: {:?} matched!", d.file_name().into_string().unwrap());
         let mut f = File::open(d.path()).unwrap();
         let mut s = String::new();
         match f.read_to_string(&mut s) {
@@ -18,7 +21,6 @@ fn main() {
             },
             _ => (),
         }
-        println!("{:?} matched", d.file_name().into_string());
     }) {
       Ok(_) => println!("done"),
       Err(e) => panic!(e)
